fix(opencli): copy adapter runtime into desktop build
