+++ packages/desktop-electron/electron-builder.config.ts
@@ -1,5 +1,4 @@
 import { execFile } from "node:child_process"
-import { existsSync, readFileSync } from "node:fs"
 import { mkdir, writeFile } from "node:fs/promises"
 import { createRequire } from "node:module"
 import path from "node:path"
@@ -8,6 +7,9 @@ import { promisify } from "node:util"
 
 import type { Configuration } from "electron-builder"
 import { writeAppUpdateConfig, type GitHubPublishConfig } from "./scripts/write-app-update-config"
+import { openCliRuntimeFileSets, openCliRuntimePackageNames } from "./opencli-runtime"
+
+export { openCliRuntimeFileSets, openCliRuntimePackageNames } from "./opencli-runtime"
 
 const execFileAsync = promisify(execFile)
 const rootDir = path.resolve(path.dirname(fileURLToPath(import.meta.url)), "../..")
@@ -16,10 +18,6 @@ const requireFromOpencode = createRequire(path.join(rootDir, "packages", "openco
 const opencodePackage = requireFromOpencode("./package.json") as {
   dependencies?: Record<string, string>
   devDependencies?: Record<string, string>
-}
-type PackageJson = {
-  name?: string
-  dependencies?: Record<string, string>
 }
 type Channel = "dev" | "beta" | "prod"
 const localizedMacDisplayNameByChannel: Record<Channel, string> = {
@@ -86,76 +84,6 @@ export function nativeWatcherFileSets() {
     from: nativeWatcherPackageDir(packageName),
     to: path.join("node_modules", ...packageName.split("/")),
     filter: ["**/*"],
-  }))
-}
-
-function readPackageJson(packageDir: string): PackageJson {
-  return JSON.parse(readFileSync(path.join(packageDir, "package.json"), "utf8")) as PackageJson
-}
-
-function packageRootFromResolvedEntry(packageName: string, resolvedEntry: string) {
-  let dir = path.dirname(resolvedEntry)
-  while (true) {
-    const packageJsonPath = path.join(dir, "package.json")
-    if (existsSync(packageJsonPath)) {
-      const json = JSON.parse(readFileSync(packageJsonPath, "utf8")) as PackageJson
-      if (json.name === packageName) return dir
-    }
-
-    const parent = path.dirname(dir)
-    if (parent === dir) {
-      throw new Error(`Could not find package root for ${packageName} from ${resolvedEntry}`)
-    }
-    dir = parent
-  }
-}
-
-function resolvePackageRoot(packageName: string, issuerPackageDir: string) {
-  const resolver = createRequire(path.join(issuerPackageDir, "package.json"))
-  const resolvedEntry = resolver.resolve(packageName)
-  if (path.isAbsolute(resolvedEntry)) return packageRootFromResolvedEntry(packageName, resolvedEntry)
-  // Bun resolves its built-in `undici` compatibility entry to the bare
-  // specifier. The installed package still exposes package.json in this layout,
-  // so use it only as a narrow fallback for non-path results.
-  return path.dirname(resolver.resolve(`${packageName}/package.json`))
-}
-
-function openCliRuntimePackages() {
-  const packages: Array<{ name: string; dir: string; json: PackageJson }> = []
-  const seen = new Map<string, string>()
-
-  function visit(packageName: string, issuerPackageDir: string) {
-    const dir = resolvePackageRoot(packageName, issuerPackageDir)
-    const previous = seen.get(packageName)
-    if (previous) {
-      if (previous !== dir) throw new Error(`Multiple runtime copies found for ${packageName}: ${previous}, ${dir}`)
-      return
-    }
-
-    const json = readPackageJson(dir)
-    seen.set(packageName, dir)
-    packages.push({ name: packageName, dir, json })
-    for (const dependency of Object.keys(json.dependencies ?? {}).sort()) {
-      visit(dependency, dir)
-    }
-  }
-
-  visit("@jackwener/opencli", path.join(rootDir, "packages", "opencode"))
-  return packages.sort((a, b) => a.name.localeCompare(b.name))
-}
-
-export function openCliRuntimePackageNames() {
-  return openCliRuntimePackages().map((pkg) => pkg.name)
-}
-
-export function openCliRuntimeFileSets() {
-  return openCliRuntimePackages().map((pkg) => ({
-    from: pkg.dir,
-    to: path.join("node_modules", ...pkg.name.split("/")),
-    filter:
-      pkg.name === "@jackwener/opencli"
-        ? ["package.json", "README.md", "LICENSE", "cli-manifest.json", "clis/**/*", "dist/src/**/*"]
-        : ["**/*"],
   }))
 }
